+++ autd3-driver/src/datagram/modulation/radiation_pressure.rs
@@ -51,11 +51,11 @@ mod tests {
 
     use super::{super::tests::TestModulation, *};
 
-    use crate::{defined::FREQ_40K, geometry::tests::create_geometry};
+    use crate::{defined::FREQ_40K, freq::kHz, geometry::tests::create_geometry};
 
     #[rstest::rstest]
     #[test]
-    #[case::freq_4k(SamplingConfig::FREQ_4K_HZ)]
+    #[case::freq_4k(SamplingConfig::Freq(4 * kHz))]
     #[case::disable(SamplingConfig::DISABLE)]
     fn test_sampling_config(#[case] config: SamplingConfig) {
         assert_eq!(
@@ -86,7 +86,7 @@ mod tests {
             )])),
             TestModulation {
                 buf: buf.clone(),
-                config: SamplingConfig::FREQ_4K_HZ,
+                config: SamplingConfig::Freq(4 * kHz),
                 loop_behavior: LoopBehavior::infinite(),
             }
             .with_radiation_pressure()