+++ autd3/src/modulation/square.rs
@@ -1,12 +1,12 @@
-use autd3_driver::derive::*;
+use autd3_driver::{derive::*, freq::FreqFloat};
 
-use super::sampling_mode::{ExactFreq, NearestFreq, SamplingMode};
+use super::sampling_mode::{ExactFreqFloat, NearestFreq, SamplingMode, SamplingModeInference};
 
 /// Square wave modulation
 #[derive(Modulation, Clone, PartialEq, Debug, Builder)]
 pub struct Square<S: SamplingMode> {
     #[get]
-    freq: f64,
+    freq: S::T,
     #[getset]
     low: u8,
     #[getset]
@@ -18,19 +18,9 @@ pub struct Square<S: SamplingMode> {
     __phantom: std::marker::PhantomData<S>,
 }
 
-impl Square<ExactFreq> {
-    pub const fn new(freq: f64) -> Self {
-        Self::with_freq_exact(freq)
-    }
-
-    /// constructor.
-    ///
-    /// # Arguments
-    ///
-    /// * `freq` - Frequency of the square wave \[Hz\]
-    ///
-    pub const fn with_freq_exact(freq: f64) -> Self {
-        Self {
+impl Square<ExactFreqFloat> {
+    pub const fn new<S: SamplingModeInference>(freq: S) -> Square<S::T> {
+        Square {
             freq,
             low: u8::MIN,
             high: u8::MAX,
@@ -45,9 +35,9 @@ impl Square<ExactFreq> {
     ///
     /// # Arguments
     ///
-    /// * `freq` - Frequency of the square wave \[Hz\]
+    /// * `freq` - Frequency of the square wave
     ///
-    pub const fn with_freq_nearest(freq: f64) -> Square<NearestFreq> {
+    pub const fn with_freq_nearest(freq: FreqFloat) -> Square<NearestFreq> {
         Square {
             freq,
             low: u8::MIN,
@@ -62,28 +52,13 @@ impl Square<ExactFreq> {
 
 impl<S: SamplingMode> Modulation for Square<S> {
     fn calc(&self, geometry: &Geometry) -> Result<HashMap<usize, Vec<u8>>, AUTDInternalError> {
-        if self.freq < 0. {
-            return Err(AUTDInternalError::ModulationError(format!(
-                "Frequency ({}Hz) must be positive",
-                self.freq
-            )));
-        }
-
         if !(0.0..=1.0).contains(&self.duty) {
             return Err(AUTDInternalError::ModulationError(
                 "duty must be in range from 0 to 1".to_string(),
             ));
         }
-
         Self::transform(geometry, |dev| {
             let (n, rep) = S::validate(self.freq, self.config, dev.ultrasound_freq())?;
-            if self.freq >= self.config.freq(dev.ultrasound_freq())? / 2. {
-                return Err(AUTDInternalError::ModulationError(format!(
-                    "Frequency ({}Hz) is equal to or greater than the Nyquist frequency ({}Hz)",
-                    self.freq,
-                    self.config.freq(dev.ultrasound_freq())? / 2.
-                )));
-            }
             Ok((0..rep)
                 .map(|i| (n + i) / rep)
                 .flat_map(|size| {
@@ -99,6 +74,8 @@ impl<S: SamplingMode> Modulation for Square<S> {
 
 #[cfg(test)]
 mod tests {
+    use autd3_driver::freq::Hz;
+
     use crate::tests::create_geometry;
 
     use super::*;
@@ -112,11 +89,24 @@ mod tests {
             0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 255, 255, 255, 255, 255, 255, 255,
             255, 255, 255, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
         ]),
-        150.
+        150.*Hz
+    )]
+    #[case(
+        Ok(vec![
+            255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 0, 0, 0, 0, 0, 0, 0,
+            0, 0, 0, 0, 0, 0, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 0,
+            0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 255, 255, 255, 255, 255, 255, 255, 255, 255,
+            255, 255, 255, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+        ]),
+        150*Hz
     )]
     #[case(
         Ok(vec![255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]),
-        200.
+        200.*Hz
+    )]
+    #[case(
+        Ok(vec![255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]),
+        200*Hz
     )]
     #[case(
         Ok(vec![
@@ -127,29 +117,39 @@ mod tests {
             255, 255, 0, 0, 0, 255, 255, 0, 0, 0, 255, 255, 0, 0, 0, 255, 255, 0, 0, 0, 255, 255,
             255, 0, 0, 0, 255, 255, 255, 0, 0, 0, 255, 255, 255, 0, 0, 0
         ]),
-        781.25
+        781.25*Hz
     )]
     #[case(
-        Err(AUTDInternalError::ModulationError("Frequency (150.01Hz) cannot be output with the sampling config (Division(5120)).".to_owned())),
-        150.01
+        Err(AUTDInternalError::ModulationError("Frequency (150.01 Hz) cannot be output with the sampling config (Division(5120)).".to_owned())),
+        150.01*Hz
     )]
     #[case(
-        Err(AUTDInternalError::ModulationError("Frequency (2000Hz) is equal to or greater than the Nyquist frequency (2000Hz)".to_owned())),
-        2000.
+        Err(AUTDInternalError::ModulationError("Frequency (2000 Hz) is equal to or greater than the Nyquist frequency (2000 Hz)".to_owned())),
+        2000.*Hz
     )]
     #[case(
-        Err(AUTDInternalError::ModulationError("Frequency (4000Hz) is equal to or greater than the Nyquist frequency (2000Hz)".to_owned())),
-        4000.
+        Err(AUTDInternalError::ModulationError("Frequency (2000 Hz) is equal to or greater than the Nyquist frequency (2000 Hz)".to_owned())),
+        2000*Hz
     )]
-    fn with_freq_exact(#[case] expect: Result<Vec<u8>, AUTDInternalError>, #[case] freq: f64) {
+    #[case(
+        Err(AUTDInternalError::ModulationError("Frequency (4000 Hz) is equal to or greater than the Nyquist frequency (2000 Hz)".to_owned())),
+        4000.*Hz
+    )]
+    #[case(
+        Err(AUTDInternalError::ModulationError("Frequency (4000 Hz) is equal to or greater than the Nyquist frequency (2000 Hz)".to_owned())),
+        4000*Hz
+    )]
+    fn with_freq_float_exact(
+        #[case] expect: Result<Vec<u8>, AUTDInternalError>,
+        #[case] freq: impl SamplingModeInference,
+    ) {
         let geometry = create_geometry(1);
-        let m = Square::with_freq_exact(freq);
+        let m = Square::new(freq);
         assert_eq!(freq, m.freq());
         assert_eq!(u8::MIN, m.low());
         assert_eq!(u8::MAX, m.high());
         assert_eq!(0.5, m.duty());
         assert_eq!(SamplingConfig::Division(5120), m.sampling_config());
-
         assert_eq!(expect.map(|b| HashMap::from([(0, b)])), m.calc(&geometry));
     }
 
@@ -160,21 +160,28 @@ mod tests {
             255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 0, 0, 0, 0, 0, 0, 0,
             0, 0, 0, 0, 0, 0, 0,
         ]),
-        150.
+        150.*Hz
     )]
     #[case(
         Ok(vec![255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]),
-        200.
+        200.*Hz
     )]
     #[case(
-        Err(AUTDInternalError::ModulationError("Frequency (2000Hz) is equal to or greater than the Nyquist frequency (2000Hz)".to_owned())),
-        2000.
+        Err(AUTDInternalError::ModulationError("Frequency (2000 Hz) is equal to or greater than the Nyquist frequency (2000 Hz)".to_owned())),
+        2000.*Hz
     )]
     #[case(
-        Err(AUTDInternalError::ModulationError("Frequency (4000Hz) is equal to or greater than the Nyquist frequency (2000Hz)".to_owned())),
-        4000.
+        Err(AUTDInternalError::ModulationError("Frequency (4000 Hz) is equal to or greater than the Nyquist frequency (2000 Hz)".to_owned())),
+        4000.*Hz
     )]
-    fn with_freq_nearest(#[case] expect: Result<Vec<u8>, AUTDInternalError>, #[case] freq: f64) {
+    #[case(
+        Err(AUTDInternalError::ModulationError("Frequency (-0.1 Hz) must be positive".to_owned())),
+        -0.1*Hz
+    )]
+    fn with_freq_nearest(
+        #[case] expect: Result<Vec<u8>, AUTDInternalError>,
+        #[case] freq: FreqFloat,
+    ) {
         let geometry = create_geometry(1);
         let m = Square::with_freq_nearest(freq);
         assert_eq!(freq, m.freq());
@@ -187,20 +194,9 @@ mod tests {
     }
 
     #[test]
-    fn freq_must_be_positive() {
-        let geometry = create_geometry(1);
-        assert_eq!(
-            Err(AUTDInternalError::ModulationError(
-                "Frequency (-0.1Hz) must be positive".to_string()
-            )),
-            Square::with_freq_nearest(-0.1).calc(&geometry)
-        );
-    }
-
-    #[test]
     fn with_low() -> anyhow::Result<()> {
         let geometry = create_geometry(1);
-        let m = Square::new(150.).with_low(u8::MAX);
+        let m = Square::new(150. * Hz).with_low(u8::MAX);
         assert_eq!(u8::MAX, m.low());
         assert!(m
             .calc(&geometry)?
@@ -213,7 +209,7 @@ mod tests {
     #[test]
     fn with_high() -> anyhow::Result<()> {
         let geometry = create_geometry(1);
-        let m = Square::new(150.).with_high(u8::MIN);
+        let m = Square::new(150. * Hz).with_high(u8::MIN);
         assert_eq!(u8::MIN, m.high());
         assert!(m
             .calc(&geometry)?
@@ -226,14 +222,14 @@ mod tests {
     #[test]
     fn with_duty() -> anyhow::Result<()> {
         let geometry = create_geometry(1);
-        let m = Square::new(150.).with_duty(0.0);
+        let m = Square::new(150. * Hz).with_duty(0.0);
         assert_eq!(m.duty(), 0.0);
         assert!(m
             .calc(&geometry)?
             .values()
             .all(|b| b.iter().all(|&x| x == u8::MIN)));
 
-        let m = Square::new(150.).with_duty(1.0);
+        let m = Square::new(150. * Hz).with_duty(1.0);
         assert_eq!(m.duty(), 1.0);
         assert!(m
             .calc(&geometry)?
@@ -250,20 +246,20 @@ mod tests {
             Err(AUTDInternalError::ModulationError(
                 "duty must be in range from 0 to 1".to_string()
             )),
-            Square::new(150.).with_duty(-0.1).calc(&geometry)
+            Square::new(150. * Hz).with_duty(-0.1).calc(&geometry)
         );
 
         assert_eq!(
             Err(AUTDInternalError::ModulationError(
                 "duty must be in range from 0 to 1".to_string()
             )),
-            Square::new(150.).with_duty(1.1).calc(&geometry)
+            Square::new(150. * Hz).with_duty(1.1).calc(&geometry)
         );
     }
 
     #[test]
     fn test_square_derive() {
-        let m = Square::new(150.);
+        let m = Square::new(150. * Hz);
         assert_eq!(m, m.clone());
     }
 }