+++ examples/src/tests/holo.rs
@@ -10,7 +10,7 @@ use std::{
 pub async fn holo(autd: &mut Controller<impl Link>) -> anyhow::Result<bool> {
     autd.send(Silencer::default()).await?;
 
-    let m = Sine::new(150.);
+    let m = Sine::new(150. * Hz);
 
     let center = autd.geometry.center() + Vector3::new(0., 0., 150.0 * MILLIMETER);
     let p = Vector3::new(30. * MILLIMETER, 0., 0.);